Initial work
Currently, this project scans Skyrim.esm (in its install directory on my personal computer) and loads just enough record data to be able to enumerate all quest forms in the file.

Eventual goals are to build a GUI for this using Qt, and use it as the basis of a standalone dialogue editor (since the CK's UI is buggy, and xEdit doesn't *quite* offer what I need for a specific project). Improve support for the ESP format (e.g. we don't handle what UESP calls "lstrings" properly yet).

Immediate goals:

Merge esp_istream and TESPlugin into a single class.

Rethink the refcounting system since loaded_form_ptr isn't working out the way we wanted it to. Among other things, although we can access TESForm members through it, (TESQuest*)my_ptr tries and fails to cast the smart pointer rather than the underlying loaded form. We can't just use a shared_ptr because we need to be able to prevent the loaded form data from unloading if there are unsaved changes, even if the refcount is zero. Ultimately the goal is to keep form data in memory only as long as we need to.

(Hm... We could use shared_ptr and then just have an object that exists solely to keep references to forms that have been modified. Ugly as sin, but doable.)