FormStubHeap
Performance testing revealed that FormStubHeap was slowing down the load process significantly (30ms with the default heap; 100ms with the custom one). After revising it to use a custom, simplified bitset class instead of std::bitset, there is no slowdown.

Gonna see if the slowdown was due to template-related nightmares or due to the lack of an equivalent to cobb::bitset::find_first_clear.