Redesigned smart pointers for loaded form data
Redesigned the loaded_form_ptr class; it's now templated, which will make it easier to cast from a pointer-to-TESForm to a pointer-to-subclass.

We now load more form types.

Form type numbers are now in an enum instead of being magic numbers.

Started work on (extremely incomplete) struct definitions related to CTDA.

Wrote a class for loading NPC_ data into memory. We don't need to load all data for the purposes of the dialogue editor. (In fact, it may be better if we give FormStub the ability to hold editor IDs and then just don't even load NPC_ data into memory beyond that. We could do this without enlarging FormStub if we cannibalize some refcount bits for flags (including the "edited" flag) and then replace the (edited) byte with a const char* pointer.)