Prep work for compressed records
Records are now read using a common interface, TESPluginRecordBuffer, which loads record data into a buffer regardless of whether or not it's compressed. The goal was to make it so that outside code doesn't have to care whether the data is compressed, calling the same "read" functions either way.

This is all very ugly and I hate it, but rewriting and refactoring isn't even going to be an option until I figure out how to get zlib to work.