Refactored loading; added support for nested GRUPs
Previously, TESPluginFile held parser state for the current top-level group, the current record, and the current subrecord, and there existed two structs (TESPluginRecord and TESPluginSubrecord) that served as interfaces to a TESPluginFile interface: they were little more than a TESPluginFile pointer with some member functions, to allow the code for loading forms into memory to read the contents of a given record or subrecord.

Some refactoring has been done: TESPluginGroup, TESPluginRecord, and TESPluginSubrecord are now whole structs (rather than fancy pointers) that store the parser state formerly kept on anonymous structs on TESPluginFile. Instances of these three structs exist as members of TESPluginFile. Separating things out makes it a little easier to do things like keeping track of the state of multiple groups. Speaking of which...

We now handle nested GRUPs -- properly, I hope, but more form-specific work needs to be done to be sure of that.

TODO: FormStub needs to be able to hold form-type-specific data. A DIAL needs to know its parent INFO; a CELL needs to know its parent WRLD and its containing block and sub-block; a REFR needs to know its parent CELL. These relationships don't exist in the records themselves; they are entirely the product of GRUP relationships and so must be resolved at parse time. (We'll also need to set up some functions for the inverse: the code that loads a DIAL into memory must be able to ask the program to load the GRUP immediately following the DIAL record -- and the INFO records in that GRUP.)