Fixed handling of nested groups
Fixed an issue where we would fail to detect the end of a nested group if that group was followed by a record; this was causing things to break immediately upon encountering a DIAL/INFO group. We should now be handling nested groups properly (though we won't know for absolute certain until we test a WRLD/B/SB/CELL/REFR group).

Improved documentation comments in memory.h.