Update formstub.h
Changed the block size for FormStubHeap. This significantly reduces the amount of memory wasted on heap-related metadata, and shaves roughly half a second off of the load time.

For reference, the load time with this change is about 2.3 seconds; this is when creating FormStubs for all form types that the program isn't programmed to skip AND all forms in all nested GRUPs regardless of type -- significantly more stubs than past tests (which completed load in 0.3 seconds), I believe.