We now load all form types; threading/memory experiments
Removed the only-for-testing optimization of skipping top-level groups outside of a narrow few; we now load ALL forms in the entire Skyrim.esm file.

Enlarged the FormStubHeap block size again, from 400 (which gives us a load time of 84 seconds for the entire ESM file) to 16000 (which gives us a load time of 48 seconds). Block size increases have diminishing returns; the only way to get any real performance improvements past this point is to multi-thread file loading.

Additionally, I've started work on resources that I'll be needing to allow the multi-threaded loading of files. The as-yet-untested cobb::multithreaded_block_allocator maintains multiple lists of blocks of memory, and allows a thread to take ownership of an entire list of blocks; this can be used to minimize the amount of time threads spend waiting on each other when allocating large numbers of objects. The new allocator depends on cobb::thread, a class that represents the state of a thread (and whose creation corresponds with the creation of a thread); cobb::thread automates the process of letting some system know when a thread has terminated: a thread must still identify itself to the other system (by passing the cobb::thread instance, which it has access to) but the cobb::thread instance does the work of tracking when the owned thread has terminated. This is something that you could *almost* do with std::async and std::future (i.e. async runs the code and the allocator holds futures), except that I don't think the code you run with std::async can get access to its own std::future (if the code that spawns a thread is responsible for giving that thread to the allocator, then you have a race condition; if the thread tries to allocate something before or while it's given to the allocator, then things could get messy).