Threading fixes
Threading fixes, and retyped some of the ideas I brainstormed in the last commit description in the documentation comments for the multi-threaded block allocator. The more I think about this, the more I feel that yeah, I think it actually may be workable to use more of the STL and make the code more portable.