form stub map
I've abandoned my intentions of creating a lock-free red-black tree, because the only pseudocode I could find was... well, completely terrible, to be blunt. Just unusably bad. Variable declarations are frequently omitted; argument-passing is frequently omitted; the authors refused to offer even a basic pseudocode struct definition where needed, but also avoided actually describing the struct and its contents using prose...

I'm currently trying a non-parallel WAVL tree, which (if I get it working) I'll just slap a lock onto. The only reason I'm not falling back to std::map with a lock just yet is because I want to be able to iterate over subtrees e.g. split the tree into eighths and have eight threads process everything in the tree. But if this, too, proves to not be worth the trouble then I'll just use a central std::map with a lock and have the loader-threads insert into it (currently each thread has its own map and we burn three seconds collating all of the maps after the threads join).