Tests for error messages; loading fixes
Fixed our cyclical dependency check. It failed because it tried to compare TESPluginHeader instances, but if we encounter the same file twice, we'll have created a new TESPluginHeader for it; we now store and compare filenames.

Fixed a mistake in TESPluginBaseReader::nextSubrecord which prevented us from reading subrecords from a record at the very end of the file. When we read subrecords, we read them from an in-memory buffer containing the entire containing record's body; this means that the file's current read position is at the end of the record. The code for nextSubrecord had leftover safety checks from when we were reading directly from the file, including an EOF check; this was causing a false-positive and preventing us from reading subrecords in the last record in the file.

TESPluginBaseReader::nextSubrecord now logs an error if we encounter an invalid XXXX subrecord (a length extension subrecord). In order to log the filename properly, we store a const char* on the base reader (somewhat unfortunate; it's redundant given the std::string on TESPluginFile, which subclasses the base reader -- ugly!).

TESPluginHeader::load now properly closes its file handle if it aborts early due to an error.

LoadOrder::loadQueuedFiles ensures that the base path contains a trailing slash. We should probably do this elsewhere, though. The function also now properly returns false if an error occurred somewhere during multi-threaded loading of a file.

LoadOrder::reset now properly deletes all TESPluginHeader instances in the load order.

I hand-created a bunch of broken ESPs in a hex editor, to test the program's error messages.