Condition progress; plans
Added a function to print condition arguments, though it's very basic and incomplete.

Started work on something experimental: singletons that act as type definitions for condition argument types, providing a uniform interface through which we can serialize argument values to strings, loop over all possible values in enum types, and so on.

Noted some near-future plans in the comments in main.cpp.