Fixed loading of Papyrus fragments; changed how we do scoped enums
Fixed the loading of Papyrus fragments: we now handle the case of there being no fragment data.

TESPluginFile and TESPluginHeader now force the Master and Light flags as appropriate for the file's extension; this behavior is consistent with the game.

Fixed cobb::bytecount_to_int_t.

Changed some enums to scoped enums.

Changed how we do scoped enums. The previous namespace approach was vulnerable to conflicts between different enums' value names, because C is a dumpster fire and so is half of its feature set. The new approach uses preprocessor macros to guarantee a unique namespace name; unfortunately, it also causes IntelliSense to crash and burn when dealing with large enums (so naturally, that only became evident when I actually put it into practice). IntelliSense displays huge swaths of these enums as syntax errors even though they're valid and the compiler is fine with them; presumably this will also break auto-completion and so on.