Critical fixes to form loading; continued work on Use Info
Modified TESPluginFile and FormStub to allow you to specify your own TESPluginBaseReader (actually, TESPluginFileView, a subclass created specifically for this) to load a record's contents into; this makes it possible for multiple threads to independently read records from the same file at the same time, which we needed for Use Info.

In addition to storing forms per form type, LoadOrder now also stores a single map of all forms. LoadOrder::getForm and LoadOrder::getFormOfProbableType have been updated to leverage this. We needed this change so that we could split forms across multiple threads when generating Use Info more evenly than would be possible by assigning whole form types to threads.

Fixed multiple mistakes in PapyrusScriptData::generateUseInfo and its helper function.

Fixed TESPluginFile::load: we now load worldspace forms, and we now have a loader for the contents of worldspaces' persistent cells.

FormStub::add_outbound_reference currently forces a breakpoint when a form is found to reference a non-existent form. We need this for debugging to detect cases where our loader skipped over forms for some reason (I'm still tracking down and fixing issues like this), but we'll want to remove it for release so that we don't choke on hand-edited user plugins (actually, we'll want to remember the dangling form ID so that we know not to create a form with that ID; we should implement that later).