Beginning to wrap up the basics for Use Info
Removed lots of debugging and benchmarking code. Added ObScript subrecords to LoadedForms::Quest::generateUseInfo's list of signatures to ignore when compiled for debug.

Currently we generate Use Info from quests only. When the program is compiled for Release, and run with a load order of Skyrim, Update, Dawnguard, and Dragonborn, it takes about a second to generate outbound connections between forms and some 50ms to generate inbound connections among forms. Of course, these figures aren't representative; as we add code for more forms' use info, we can expect things to get quite a bit slower.