Use Info progress
Adjusted C++ file output paths to prevent the linker from choking on files with different paths but the same name. I'm mystified that this isn't a default configuration in The Year Of Our Divines 4E 201.

We are now capable of loading FormLists and VoiceTypes, and we generate Use Info for those as appropriate as well as for Containers and MagicEffects.

Moved handling for COCT/CNTO/COED subrecords into a dedicated struct ContainerData, since those appear in multiple record types. Created a dedicated struct FormModel for MODL/MODT/MODS.

FormStubs now use factory functions to load form data; this means that formstub.cpp won't need to include the headers for every single loaded form class.