Memory management
Block allocation is now disabled for Use Info data and for the contents of form-ID-to-FormStub maps. This is significantly faster; apparently, as long as we block-allocate the FormStubs themselves, the standard heap is fast enough for everything else.

This is the case, by the way, even after optimizing our block allocator to better handle huge block sizes. Blocks now keep track of their first free slot and their number of remaining slots, so we can more quickly skip full blocks (and find free slots) when allocating. The allocator is still extremely slow when used for binary tree nodes.

Added cobb::bitset::find_first_clear_from(index).