Renaming this project
SkyEdit was a working title, and unfortunately, it seems to have already been taken: the UESP has a 2012-era effort at creating an open-source Creation Kit equivalent.

As such, we're going with DovahKit. Quite proud of the name, personally. This is all the files renamed and relocated; I'll be changing the name of the repo itself next.