Working on a block-allocator that may satisfy the Allocator named requirement
I'm working on a block allocator that satisfies the constraints of Allocator, including the ability to rebind it to arbitrary types without breaking it. I found a Qt blog article in which they used a static thread_local unique_ptr member on a heap class to get around the need to manually register threads. A thread_local variable has to be global or effectively global (static members on a class count); the compiler makes it so that when a thread tries to access the variable, a copy is constructed there and then just for that thread, and then destroyed when the thread dies.