cobb::multiheap
I now have a multi-threaded block allocator/heap that no longer needs to be manually told when threads spawn and die. In tests, using it with std::map is still significantly slower (ESPECIALLY when mass-freeing) than a std::unordered_map with no cusotm allocator, but still -- we have a block allocator and it's compatible with the Allocator named requirement and STL containers that use it.