Started work on "active file" functionality
It's not tested yet, but you can specify an Active File. The LoadOrder singleton will remember which file that is, and keep track of its forms both in the normal form maps and separately. The active file must be at the end of the computed load order (we'll move it there if possible and fail if not), and (by definition) it cannot be the master of any other files in the load order.

Remembered to have ACTI (Activator) generate use info for its model.