Object Window work
The object window's table view is now functional; it can be filtered using the form type tree view, though not yet using the textbox for editor and form IDs.

The table view can be sorted, though sorting by editor ID is fairly slow and causes a hitch for a second or two.