Loaded form work; Shout editing UI
Finished up most of the existing form types' load methods.

Built an editing dialog box for Shouts (accessible by double-clicking a Shout in the Object Window), as part of an attempt at testing how well we can set up editing for form types in general. Currently, comboxes that list all forms of a given type incur a heavy performance hit, and I'm not sure of the best way to mitigate that.

Fixes to FormTableModelItem.