Misc fixes; added form types: Color, Word of Power
Added form types: Color, Word of Power.

Added editing dialogs for form types: Color, Voicetype, Word of Power.

Made a minor tweak to FormsOfTypeCombobox to improve performance. Populating it still lags in Debug, but is almost unnoticeable in Release.

Fixed the Voicetype class so that its unique flags don't accidentally shadow the form flags.

Fixed a mistake in the Object Window that caused it to list Actors (as in, placed characters in the world) rather than ActorBases.