File load dialog work; fix for zero-length recordsq
Zero-length records no longer cause loading to fail. Previously, they would cause loading to fail not because they read as invalid, but because they were indistinguishable from memory allocation failures: cobb::generic_buffer doesn't allocate any memory if you tell it to allocate zero bytes, and it also doesn't allocate any memory if there isn't enough room for what you're asking for. Test-case for this fix was a zero-length NAVM in Dawnguard.esm.

DovahKitCore now supports performing a file load asynchronously, so as not to leave the full program unresponsive.

The file-open dialog now supports setting an active file and can now load the selected files.

DovahKitCore can now supply performance information after a successful file load. Currently, the main window catches this and displays the information in the status bar.

Fixed cobb::windows_registry::get_string_value: it now trims off spare null characters on the end of the output string. This is needed because the string must be resize()d, not reserve()d, to make room for the output, and so it'll have a ton of null space at the end if the output is smaller than the reserved space.

Wrote an extensive to-do list for the near future.

Edited the repo readme to explain the program's name.