Taskbar progress bar
DovahKit now shows the progress for loading files in its taskbar button. We can show the progress for creating form_stubs and for generating outbound use info; inbound use info, however, is not so straightforward, at least at this time. The underlying behavior involves the load window polling the current load order to ask for a progress measurement (which isn't thread-safe except where needed to avoid crashes, and therefore isn't 100% precise, but it's not *noticeably* imprecise).

To facilitate the above change, some structural changes were made to file_load_order with respect to how it keeps track of the threads for building outbound use info.