Basic file writing test works properly
The file_load_order class now properly adds forms that are defined in a non-active file but overridden in the active file to the active-file-forms maps.

The file_reader class now stores the file author and description as std::string, and these are now properly written to the file.

Various fixes to file writing, including properly byteswapping signatures when writing them.

Redesigned cobb::generic_buffer to use realloc under the hood and to have function names that better match STL containers.

We can now correctly re-save the Skyrim Classic version of my "Bony Footsteps for Dragonbone Boots" mod, which just overrides an ArmorAddon to use a new Footstep Set.