File save dialog
Added a dialog box for saving the active file. It's still a work in progress; it needs to display dependencies and pop a confirmation prompt before overwriting files.

Fixed a bug where if there were no forms of a given type and you tried to filter the object window to that type, a crash would occur.