Wrote the rest of the save process
Wrote the rest of the save process -- specifically, the file swap.

Previous implementations of file saving relied on iterating through all top-level active file forms by type and writing them out; however, this would miss an edited child form somewhere within an unedited, non-active-file top-level form. I've solved this by making it possible to loop over all top-level forms that need to be saved, either because they themselves are edited or were defined in the active file, or because any of their descendants were.

Previous implementations of file saving would write all of a top-level form's child forms to the file, even if the top-level form was an override and a child wasn't overridden or edited. This should now be fixed.