Started work on a Cell View window
Started work on a Cell View window. It supports showing a list of all cells in a given worldspace (or all interior cells), and showing a list of all references in the selected cell.

FormsOfTypeCombobox now allows us to use text other than " NONE" for the none item, e.g. " Interiors" for the Cell View's worldspace combobox. Unfortunately, sorting is still very basic and relies on a leading space as a cheap trick to place the none item above all others.

Fixed dovah::form_stub_helpers::for_each_child_form: parent/child references are from the child to the parent.