Cell View improvements
The Cell View window now shows unnamed worldspace cells as "Unnamed Cell" in italics. They are also sorted at the end of the list.

The jump-to-grid spinboxes in the Cell View window no longer expand.

The Cell View window now shows the name of the selected cell above the reference list. If the selected cell is an exterior cell, then its grid coordinates are displayed as well.