Started work on a cell-editing dialog
Started work on a dialog that can be used to edit cell properties. Currently, it's capable of displaying almost all of them. Music type editing is incorrect (we need to be able to specify "undefined" a.k.a. "default" alongside "none"), the interior-only tabs aren't hidden when working with an exterior cell, and there's no ability to commit changes to the form.