Rough draft of a REFR editing dialog
I created a rough draft of a dialog for editing REFR properties. This one should be a lot less painful to use than the CK one, owing to the use of a navigation sidebar rather than a tabbar with so many tabs (and in such a narrow area) that you have to button-mash to scroll it.