TESFilePicker; string extra-data load fix; cell XWEM
Added a TESFilePicker widget. For now, it's just a simple line edit, but in the future, I hope to add a "browse" button that will work with both loose files and files in the loaded BSAs.

The cell UI now uses TESFilePicker to make CELL/XWEM available for editing.

Fixed a mistake that caused the string_extra_data template to shear off the last character of any given extra-data.