File-writing compression work
The (file_writer) class now stores most of its extra configuration information in a (write_config) struct, to make it easier for outside code to pass configuration information in (given that the (file_load_order) creates a (file_writer) on the stack, so outside callers can't configure it themselves).

The "save" dialog now allows users to control how records are compressed. Testing shows that the "threshold" mode works as intended.

The "save" dialog's file flag options should now be usable.

Fixed the writing of FACT/FNAM.