File writing fixes
Fixed a mistake in file_writer::get_output_position that caused us to write the record and group count incorrectly.

We no longer open a top-level CELL GRUP if no interior cells need to be saved. (Previously, we would open an empty GRUP for interior cells if there were exterior cells that needed to be saved.)

Tweaked the Bethesda-mimicking compression policy after testing with the CK.