Form creation
It is now possible to right-click in the Object Window and create a new form of whatever type you're currently filtering to.

The FormTableModel no longer uses a root node, and handles most editor signals itself. The owning widget is still allowed to dictate when and how the dataAcquireComplete signal is reacted to, though I'll probably change that.

Added a Form::setup function to allow newly-created forms to set up uses of hardcoded forms (e.g. worldspaces using DefaultWater rather than NONE as their water types).