Improved form creation
Forms are now created as a multi-stage process, making it possible to first reserve a form ID and then ask the user for further information (including giving them the opportunity to cancel creating the form), before actually committing and creating the form.

In order to still send the formCreated signal, I needed to add an "on form create" callback to file_load_order. Hacky, but it works.