Planning for form duplication
Form creation needs to be able to specify a parent form. The presence or absence of a parent form should be the deciding factor in whether a newly-created CELL is an interior or exterior.

Two Use Info flags can be merged, and a new flag must be added for dialogue system relationships that aren't strictly parent/child. Form duplication must then pay attention to these relationships, as duplicating a QUST should generally also duplicate its pseudo-child DIALs and their child INFOs.