form_duplication_request and form-duplicating UI
Implemented form_duplication_request, a means of cloning a form and all of its children.

Implemented the missing Worldspace::large_reference_t::clone_from(...) function.

Implemented a UI for duplicating forms through the Object Window. This allows us to test the duplication of single forms (which works), but at present gives us no way to test the duplication of forms with children. (We could wire a few things together for QUST, and then write the code to load DIAL and INFO; alternatively, we can set up a UI for managing WRLD and CELL and then test duplicating an interior cell.)