Misc fixes and some preparation for dialogue handling
Fixed a typo in an assert in the Use Info window.

Merged the "is reference" and "is base form" use info flags, and added a "dialogue" flag for a topic's owning quest and for the topic info linked lists. Added dialogue_form_id_t in preparation for those as well.

The form stub helper function for getting a worldspace cell by grid coordinates should no longer be able to return the persistent cell (which should always be (0, 0) and which is distinct from the actual (0, 0) cell).

Added a form stub helper function for looping over a quest's owned dialogue topics.

Duplicating a quest should now attempt to duplicate its dialogue topics as well.

A form stub's group type should now default-initialize to -1. This is needed because the group type is used to tell a worldspace's persistent cell apart from its other cells.