UI tweaks
Table views that have a form type signature column now size the column based on the width of string "XMMX" instead of "XXXX", as "M" is a wider character. I was seeing some truncation in some listings.

The Use Info dialog should no longer show uses-as-base-form when viewing use info for references, as references cannot be base forms.