Form deletion work
When a form is deleted, all other forms that are using it need to sever their outbound connections to it. Implementing this took a considerable amount of work: all forms need code to sever outbound references, which means all form components (particularly all extra-data types) need code to do the same.

Implemented (form_deletion_request), which coordinates the deletion of a "target" form and all of its child forms, along with the severing of all references to and from these forms. Testing is pending.

Discovered a flaw with form_id_t.