Prep for form deletion
Did some testing in the Creation Kit and confirmed a few things about form deletion: flagging a CELL as deleted also flags all of its child forms, individually, as deleted; and some form types don't save any subrecords when flagged as deleted.

Wrote a speculative implementation of "friendly delete" for ObjectReference.

All UI controls that listen for the dataAbandonImminent signal should now also listen for formDeletionImminent. Dialogs dedicated to a form should close if that form is deleted; listings that show a form should remove the form if it is deleted. There are two flaws here: the UI has no way of knowing whether a form is truly being deleted or merely being flagged as deleted; and there's no way to check, using a form_stub alone, whether a form is flagged as deleted, so the UI can't show a "deleted flag" indicator on forms unless it goes to the trouble of loading them into memory in full.