Form deletion confirmation dialog is done
Form deletion now has a confirmation dialog, which lists the forms to be deleted and their users. The confirmation dialog is shown by DovahKitCore, so individual UI views don't need to do it themselves.

Fixed a crash that could occur when calling the form_stub overload for form_id_t::set and friends.