BSA work and testing
Added accessors to (bsa_archive) to allow viewing error information, and to allow looping over all folders and files.

Wrote some tests to confirm that we are properly loading BSAs and can properly provide access to their contained files. DovahKit has a debug command to enumerate a BSA's contents (in the debugger's log pane), and another debug command to extract a file. Inspection in a hex editor confirms that files (whether compressed or not) are byte-identical to the originals (as extracted by community tools).