Individual form renumbering is essentially complete
Fixed an oversight where I forgot to update form_stub::formID when renumbering.

The object window now auto-suggests the form's current ID, and warns if you choose to "renumber" a form to the ID it's already using.

All that's left is to prevent the user from using IDs below xx000800, since those are reserved for hardcoded forms. (That's not accounted for by the check to see if a form is already using the desired ID, because not all of the reserved IDs are actually in use by the engine.)