Continued work for supporting ESLs
Fixed an oversight in (file_load_order::accept_form_stub) that didn't support ESL IDs.

It is now possible to toggle whether (file_load_order) is honoring the "light" flag on file headers. That way, Skyrim Classic files with bad flags won't be treated as light, and we won't reserve slot 0xFE for light files when loading for Skyrim Classic.

The load window should now allow you to view tooltips on files in the file list, in case their names are truncated.

Various adjustments to allow the user to choose whether to load files for Skyrim Classic or Skyrim Special.

Updates to the to-do list.