We can now load SSE files, including ESLs.
We can now generate form stubs for LENS and VOLI.

Made some fixes to the subrecord reader class, which could read past the end of a subrecord. It was relying on the bounds-checking in the record reader, but that only cares about record bounds, not subrecord bounds.

Based on testing the Survival, Bone Wolf, and Armored Mudcrab ESLs, we seem to be able to load ESLs properly. Of course, we can't save them (or anything that uses them as a master) properly yet.