Code to delete lost forms during a conversion between games
Modified the save process to delete forms from memory if they were not successfully written out to the file. If any such in-memory deletions fail, then the save process halts with an error, signalling to the frontend that further editing isn't possible.

When serializing a form_reference_t, the subrecord writer will now check that the referenced form can be saved to the target file. If it can't, then form ID 0 will be serialized instead.

File write warnings now use notice codes instead of a bespoke enum.