Save process cleanup: warning for imminent form loss during conversion
DovahKit now displays a very rudimentary warning if converting the active file to another game would result in forms being lost. We need to flesh out the warning -- replace the standard message box with a custom one that lists the forms to be deleted and grants access to their use info.