Adjusted max file load limit; finally fixed cobb::multiheap
The load window now supports multi-selection of files, and offers context menu options for mass (un)checking files.

When loading Skyrim Special content, DovahKit now allows you to load more than 255 total files, so long as you don't exceed the respective limits for light files (4096) and non-light files (254). I tested this by loading 284 identical copies of the Bone Wolf ESL all at once.

Found and fixed the bug in cobb::multiheap: it was a concurrency issue, where code in one thread was attempting to free a form_stub that had been allocated on one subheap (circa line 230) just as that subheap was closing down and deleting its first block (at line 209). The Bone Wolf test that I used was able to reproduce this issue every time, and eventually it occurred to me to check the call stack on the thread with the corrupt subheap, and not just the thread that actually crashed.