Started work on a log window
The (file_load_order) is now capable of retaining a list of load warnings, and of notifying the frontend whenever a new load warning is added. Currently, the only implemented warning is for override records with mismatched types.

The main window now stores its subwindows a bit differently, which makes it more convenient to implement things like re-showing a subwindow. There is now a menu that can be used to bring a subwindow to the front (and reopen it if it was closed).

Added a very basic log window.