Log window no longer shows duplicate warnings
The log window will no longer re-show duplicate file load warnings. To achieve this, it uses a custom model. This should help with warnings that are generated from on-demand form loading.

File load warnings can now specify a cause subrecord index, to distinguish between multiple subrecords that have the same signature. Of course, this would only be useful when a subrecord is expected to recur.