Rejected form stubs no longer leak; reverse-engineering re: GMSTs
We now delete form stubs from memory if the file load order fails to accept them.

I've discovered that we fail to load a GMST record from Dawnguard.esm, because we ignore load order prefixes on GMST records to avoid choking on a Skyrim.esm GMST with an invalid form ID. Further reverse-engineering has established that GMST form IDs don't need to be valid so long as they don't end up overriding a non-GMST form. Plans are now in place to handle GMSTs properly and in a manner consistent with the game engine.

I've updated my internal documentation of the form load process to describe special-case load behaviors for specific form types, as well as to describe what I was able to learn about the TOFT record.