Work on GMST support
The file_load_order and file reader have been coded to load GMSTs differently. We no longer false-positive on that one Dawnguard.esm GMST, and we still don't choke on the out-of-bounds GMST in Skyrim.esm.

Created a UI file for a game setting editor window; however, underlying code still needs to be written, so we can see whether we're actually loading GMSTs properly.