Work on GMST support
I wrote an x96dbg script to dump all game settings (and let me tell you: that was not a pleasant activity), so now we have a comprehensive list of every GMST that TESV.exe recognizes.

Boolean GMSTs are now loaded properly, and unrecognized subrecords in GMST are now logged properly.