Work on GMST support: initial code for editing
The game setting UI is now able to indicate which settings are edited in the active file.

The game setting UI now reserves enough space to show the longest defined game setting name.

Wrote code for a "game setting change request" object. Next up is the backend-to-frontend bridge for this.