Work on GMST support
Fixed a deadlock that would've occurred when editing a game setting, due to locking a non-recursive mutex multiple times in the same block.

Wrote code to bridge the backend's GMST-editing API to the frontend, and added UI code to make it work. The UI now has a button to explicitly commit changes to a game setting. Code has not yet been built or tested.