Work on GMST support: implemented saving
The file-writer class is now capable of saving game settings.

Fixed an oversight in the save code that could cause a crash if a form stub was deleted during the save process as a result of it not having been written to the saved file. The crash was caused by iterators being invalidated during the loop -- the result of attempting the deletion within the loop.

Added half a dozen more game setting descriptions.