GMST saving fixes; work on GMST renumbering
Fixes to GMST saving: we now properly update the active file's GMST form IDs when changing whether the file is flagged as light.

Initial implementation of GMST renumbering. Currently not hooked up to the frontend, and therefore not tested.

The game_setting_edit_request class no longer needs to be a friend of file_load_order, as the latter now offers an accessor through which the former's destructor can abandon a form ID reservation.