Work on singleton form support
The (file_load_order) functions for iterating over the form maps now double-check that stubs exist before sending them to the functors, just as an extra protection for cases where we null out an entry in the form maps without explicitly using std::unordered_map::erase.

I thought of a clever way to get the "canonical" form stub for a singleton form, based on the number of source file info structs, so we now use that instead of any sorting code. Much simpler.

Wrote up internal documentation for how we handle singleton forms.