Plans: dangling form references; incorrectly-referenced GMSTs
I realized that there's an issue with on-demand form loading that affects both incorrectly-referenced GMSTs (i.e. cases where a form was tampered with to refer to a GMST), and forms with outbound references to non-existent forms (i.e. cases where a form was tampered with to refer to a form ID that was never actually used by a valid form). In the former case, operations that renumber or delete a GMST form stub could leave dangling references (from unloaded forms) or pointers (from already-loaded forms). In the latter case, a new form created using the referenced-but-unoccupied form ID would retroactively be considered the referent of the referencing form, with DovahKit unable to detect the confusion.

Dealing with the GMST case is easy enough. Dealing with the case of references to non-existent forms is harder: it essentially requires that we create and maintain form stubs for these unoccupied form IDs as the references are found (which, logically, would be during the step where we generate use info).

Aside from the above, I've cleaned up my to-do list a bit.